# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Banco D$ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  balance: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export interface PublicUser {
  id: number;
  username: string;
  avatarUrl?: string | null;
}

export interface UpdateProfileRequest {
  avatarUrl?: string | null;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
  confirmNewPassword: string;
}

export interface DeleteAccountRequest {
  password: string;
}

export interface TransferRequest {
  toUsername: string;
  amount: string;
}

export interface TransferResponse {
  message: string;
  newBalance: string;
}

export interface Reserve {
  id: number;
  name: string;
  balance: string;
  userId: number;
  createdAt: string;
}

export interface CreateReserveRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
}

export interface UpdateReserveRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
}

export interface ReserveTransactionRequest {
  amount: string;
}

export interface SharedReserveMember {
  userId: number;
  username: string;
  avatarUrl?: string | null;
  joinedAt: string;
}

export interface SharedReserve {
  id: number;
  name: string;
  balance: string;
  createdByUserId: number;
  members: SharedReserveMember[];
  createdAt: string;
}

export interface CreateSharedReserveRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  memberUsernames: string[];
}

export interface AddMemberRequest {
  username: string;
}

export interface RemoveMemberResponse {
  message: string;
  votesNeeded?: number | null;
  votesRegistered?: number | null;
  removed: boolean;
}

export interface Transaction {
  id: number;
  userId: number;
  type: "ingreso" | "egreso";
  amount: string;
  counterpartUsername?: string | null;
  description: string;
  createdAt: string;
}