# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Banco D$ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reserves:
    get:
      operationId: getReserves
      tags: [reserves]
      summary: Get all personal reserves for current user
      responses:
        "200":
          description: List of reserves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reserve"
    post:
      operationId: createReserve
      tags: [reserves]
      summary: Create a personal reserve
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReserveRequest"
      responses:
        "201":
          description: Reserve created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserve"
  /reserves/{reserveId}:
    patch:
      operationId: updateReserve
      tags: [reserves]
      summary: Update reserve (rename)
      parameters:
        - name: reserveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReserveRequest"
      responses:
        "200":
          description: Reserve updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserve"
        "404":
          description: Reserve not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReserve
      tags: [reserves]
      summary: Delete a personal reserve (only if balance is zero)
      parameters:
        - name: reserveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reserve deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Reserve has non-zero balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reserves/{reserveId}/deposit:
    post:
      operationId: depositToReserve
      tags: [reserves]
      summary: Deposit money into a personal reserve
      parameters:
        - name: reserveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveTransactionRequest"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserve"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reserves/{reserveId}/withdraw:
    post:
      operationId: withdrawFromReserve
      tags: [reserves]
      summary: Withdraw money from a personal reserve
      parameters:
        - name: reserveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveTransactionRequest"
      responses:
        "200":
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserve"
        "400":
          description: Insufficient reserve balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
      required:
        - username
        - password
        - confirmPassword
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        balance:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - balance
        - createdAt
...
    Reserve:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        balance:
          type: string
        userId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - balance
        - userId
        - createdAt
    CreateReserveRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
      required:
        - name
    UpdateReserveRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
      required:
        - name
    ReserveTransactionRequest:
      type: object
      properties:
        amount:
          type: string
      required:
        - amount
...