# lib/api-zod/src/generated/types/createReserveRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Banco D$ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReserveRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
}